+++ waves/parameter_generators.py
@@ -380,30 +380,13 @@ class ParameterGenerator(ABC):
         requires:
 
         * ``self._set_hashes``: parameter set content hashes identifying rows of parameter study
+        * ``self.set_name_template``: Parameter set name template. Overridden by ``output_file_template``, if provided
 
         creates attribute:
 
         * ``self._set_names``: Dictionary mapping parameter set hash to parameter set name
         """
-        self._set_names = {}
-        for number, set_hash in enumerate(self._set_hashes):
-            template = self.set_name_template
-            self._set_names[set_hash] = template.substitute({"number": number})
-
-    def _update_set_names(self) -> None:
-        """Update the parameter set names after a parameter study dataset merge operation.
-
-        Resets attributes:
-
-        * ``self.parameter_study``
-        * ``self._set_names``
-        """
-        self._create_set_names()
-        new_set_names = set(self._set_names.values()) - set(self.parameter_study.coords[_set_coordinate_key].values)
-        null_set_names = self.parameter_study.coords[_set_coordinate_key].isnull()
-        if any(null_set_names):
-            self.parameter_study.coords[_set_coordinate_key][null_set_names] = list(new_set_names)
-        self._set_names = self.parameter_study[_set_coordinate_key].to_series().to_dict()
+        self._set_names = _create_set_names(self._set_hashes, self.set_name_template)
 
     def _create_set_names_array(self) -> xarray.DataArray:
         """Create an Xarray DataArray with the parameter set names using parameter set hashes as the coordinate
@@ -497,7 +480,10 @@ class ParameterGenerator(ABC):
             raise RuntimeError("Called without a previous parameter study")
 
         previous_parameter_study = _open_parameter_study(self.previous_parameter_study)
-        self.parameter_study = _merge_parameter_studies([previous_parameter_study, self.parameter_study])
+        self.parameter_study = _merge_parameter_studies(
+            [previous_parameter_study, self.parameter_study], self.set_name_template
+        )
+        self.parameter_study = self.parameter_study.swap_dims({_set_coordinate_key: _hash_coordinate_key})
         previous_parameter_study.close()
 
         # Recover parameter study numpy array(s) to match merged study
@@ -505,7 +491,7 @@ class ParameterGenerator(ABC):
 
         # Recalculate attributes with lengths matching the number of parameter sets
         self._set_hashes = list(self.parameter_study.coords[_hash_coordinate_key].values)
-        self._update_set_names()
+        self._set_names = self.parameter_study[_set_coordinate_key].to_series().to_dict()
         self.parameter_study = self.parameter_study.swap_dims({_hash_coordinate_key: _set_coordinate_key})
 
 
@@ -1482,13 +1468,17 @@ def _coerce_values(values: typing.Iterable, name: typing.Optional[str] = None) -
     return values_coerced
 
 
-def _merge_parameter_studies(studies: typing.List[xarray.Dataset]) -> xarray.Dataset:
-    """Merge a list of parameter studies into one study. Set hashes are not resolved by this function at the moment.
+def _merge_parameter_studies(
+    studies: typing.List[xarray.Dataset], template: typing.Optional[_utilities._AtSignTemplate] = None
+) -> xarray.Dataset:
+    """Merge a list of parameter studies into one study.
 
     Preserve the first given parameter study set name to set contents associations by dropping subsequent studies' set
     names during merge.
 
     :param studies: list of parameter study xarray Datasets where the first study is considered the 'base' study
+    :param template: parameter set naming template utilizing the '@' sign to mark substitution. If none is provided upon
+        call, fetch it from the WAVES settings.
 
     :return: parameter study xarray Dataset
 
@@ -1521,7 +1511,54 @@ def _merge_parameter_studies(studies: typing.List[xarray.Dataset]) -> xarray.Dat
         if new_dtype != old_dtype:
             study_combined[key] = study_combined[key].astype(old_dtype)
 
+    # Recalculate attributes with lengths matching the number of parameter sets
+    study_combined = _update_set_names(study_combined, template)
+    study_combined = study_combined.swap_dims({_hash_coordinate_key: _set_coordinate_key})
+
     return study_combined
+
+
+def _create_set_names(
+    set_hashes: typing.List[str], template: typing.Optional[_utilities._AtSignTemplate] = None
+) -> dict:
+    """Construct parameter set names from the set name template and number of parameter set hashes.
+
+    :param set_hashes: parameter set content hashes identifying rows of parameter study
+    :param template: parameter set naming template utilizing the '@' sign to mark substitution. If none is provided upon
+        call, fetch it from the WAVES settings.
+
+    :return: Dictionary mapping parameter set hash to parameter set name
+    """
+    if not template:
+        template = _utilities._AtSignTemplate(_settings._default_set_name_template)
+
+    set_names = {}
+    for number, set_hash in enumerate(set_hashes):
+        set_names[set_hash] = template.substitute({"number": number})
+
+    return set_names
+
+
+def _update_set_names(
+    parameter_study: xarray.Dataset, template: typing.Optional[_utilities._AtSignTemplate] = None
+) -> xarray.Dataset:
+    """Update the parameter set names after a parameter study dataset merge operation.
+
+    :param parameter_study: A :class:`ParameterGenerator` parameter study Xarray Dataset with swapped set hash and set
+        name dimensions
+    :param template: parameter set naming template utilizing the '@' sign to mark substitution. If none is provided upon
+        call, fetch it from the WAVES settings.
+
+    :return: parameter study xarray Dataset
+    """
+    set_hashes = list(parameter_study.coords[_hash_coordinate_key].values)
+    set_names = _create_set_names(set_hashes, template)
+    new_set_names = set(set_names.values()) - set(parameter_study.coords[_set_coordinate_key].values)
+    null_set_names = parameter_study.coords[_set_coordinate_key].isnull()
+    if any(null_set_names):
+        parameter_study.coords[_set_coordinate_key][null_set_names] = list(new_set_names)
+
+    return parameter_study
 
 
 _module_objects = set(globals().keys()) - _exclude_from_namespace